+++ src/main/java/com/divary/domain/image/service/ImageService.java
@@ -18,10 +18,7 @@ import javax.imageio.ImageIO;
 import java.awt.image.BufferedImage;
 import java.io.IOException;
 import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
 import java.util.*;
-import java.util.regex.Pattern;
-import java.util.regex.Matcher;
 import java.util.stream.Collectors;
 
 
@@ -32,19 +29,17 @@ import java.util.stream.Collectors;
 public class ImageService {
 
     // 상수 정의
-    private static final int MAX_FILES_PER_UPLOAD = 10; // 최대 10개 업로드 가능
-    private static final long MAX_FILE_SIZE_BYTES = 10 * 1024 * 1024; // 10MB
     private static final int TEMP_IMAGE_EXPIRY_HOURS = 24;
-    private static Pattern tempUrlPattern = null;
 
     private final ImageRepository imageRepository;
     private final ImageStorageService imageStorageService;
     private final ImagePathService imagePathService;
+    private final ImageValidationService imageValidationService;
 
     // 다중 임시 이미지 업로드
     @Transactional
     public MultipleImageUploadResponse uploadTempImages(List<MultipartFile> files, Long userId) {
-        validateMultipleFiles(files);
+        imageValidationService.validateMultipleFiles(files);
         
         List<ImageResponse> uploadedImages = new ArrayList<>();
         
@@ -69,7 +64,7 @@ public class ImageService {
             ImageMetadata metadata = extractImageMetadata(file);
             
             // temp 경로 생성
-            String tempPath = generateTempPath(userId);
+            String tempPath = imagePathService.generateTempPath(userId);
             String fileName = imageStorageService.generateUniqueFileName(file.getOriginalFilename());
             String tempS3Key = tempPath + "/" + fileName;
             
@@ -98,10 +93,68 @@ public class ImageService {
         }
     }
 
+    // 본문에서 temp 이미지 URL을 찾아서 permanent URL로 변환 (다른 서비스에서 사용해야하는 메서드)
+    @Transactional
+    public String processContentAndMigrateImages(String content, ImageType imageType, Long userId, String additionalPath) {
+        if (content == null || content.trim().isEmpty()) {
+            return content;
+        }
+
+        List<String> tempUrls = imagePathService.extractTempImageUrls(content);
+        if (tempUrls.isEmpty()) {
+            return content;
+        }
+
+        Map<String, String> urlMappings = new HashMap<>();
+        for (String tempUrl : tempUrls) {
+            try {
+                String permanentUrl = migrateTempImageByUrl(tempUrl, imageType, userId, additionalPath);
+                urlMappings.put(tempUrl, permanentUrl);
+            } catch (Exception e) {
+                log.error("이미지 이동 실패: tempUrl={}, userId={}", tempUrl, userId, e);
+            }
+        }
+
+        return imagePathService.replaceTempUrls(content, urlMappings);
+    }
+    
+    // 게시글 수정 시 삭제된 이미지 처리 (다른 서비스에서 사용해야하는 메서드)
+    @Transactional
+    public void processDeletedImagesAfterPostUpdate(ImageType imageType, Long postId, String newContent) {
+        // 현재 게시글에 연결된 이미지 목록 조회
+        List<Image> currentImages = imageRepository.findByTypeAndPostId(imageType, postId);
+
+        if (currentImages.isEmpty()) {
+            return;
+        }
+
+        // 새 컨텐츠에서 이미지 URL 추출
+        List<String> newImageUrls = imagePathService.extractAllImageUrls(newContent);
+
+        // 삭제할 이미지 찾기 (현재 DB에는 있지만 새 컨텐츠에는 없는 이미지)
+        List<Image> imagesToDelete = currentImages.stream()
+                .filter(image -> {
+                    String imageUrl = imageStorageService.generatePublicUrl(image.getS3Key());
+                    return !newImageUrls.contains(imageUrl);
+                })
+                .collect(Collectors.toList());
+
+        // 삭제 처리
+        for (Image imageToDelete : imagesToDelete) {
+            try {
+                deleteImage(imageToDelete.getId());
+                log.info("게시글 수정으로 인한 이미지 삭제: imageId={}, postId={}, imageType={}",
+                        imageToDelete.getId(), postId, imageType);
+            } catch (Exception e) {
+                log.error("이미지 삭제 실패: imageId={}, postId={}", imageToDelete.getId(), postId, e);
+            }
+        }
+    }
+    
     // 이미지 업로드 메인 로직
     @Transactional
     public ImageResponse uploadImage(ImageUploadRequest request) {
-        validateUploadRequest(request);
+        imageValidationService.validateUploadRequest(request);
         
         try {
             ImageMetadata metadata = extractImageMetadata(request.getFile());
@@ -169,37 +222,10 @@ public class ImageService {
         imageRepository.delete(image);
     }
 
-    private void validateUploadRequest(ImageUploadRequest request) {
-        // 파일 검증
-        if (request.getFile() == null || request.getFile().isEmpty()) {
-            throw new BusinessException(ErrorCode.VALIDATION_ERROR);
-        }
-        
-        // 업로드 경로 검증
-        if (request.getUploadPath() == null || request.getUploadPath().trim().isEmpty()) {
-            throw new BusinessException(ErrorCode.VALIDATION_ERROR);
-        }
-        
-        // 경로 보안 검증 (../ 등 상위 디렉토리 접근 차단)
-        if (request.getUploadPath().contains("..") || request.getUploadPath().startsWith("/")) {
-            throw new BusinessException(ErrorCode.VALIDATION_ERROR);
-        }
-        
-        // 파일 크기 검증 (10MB 제한)
-        if (request.getFile().getSize() > 10 * 1024 * 1024) {
-            throw new BusinessException(ErrorCode.VALIDATION_ERROR);
-        }
-        
-        // 파일 타입 검증
-        String contentType = request.getFile().getContentType();
-        if (contentType == null || !contentType.startsWith("image/")) {
-            throw new BusinessException(ErrorCode.VALIDATION_ERROR);
-        }
-    }
 
     // 타입별 이미지 업로드
     @Transactional
-    public ImageResponse uploadImageByType(ImageType imageType, org.springframework.web.multipart.MultipartFile file, Long userId, String additionalPath) {
+    public ImageResponse uploadImageByType(ImageType imageType, MultipartFile file, Long userId, String additionalPath) {
         // 업로드 경로 생성
         String uploadPath;
         if (imageType.name().startsWith("USER_")) {
@@ -238,52 +264,6 @@ public class ImageService {
                 .collect(Collectors.toList());
     }
 
-    // 본문에서 temp 이미지 URL을 찾아서 permanent URL로 변환
-    @Transactional
-    public String processContentAndMigrateImages(String content, ImageType imageType, Long userId, String additionalPath) {
-        if (content == null || content.trim().isEmpty()) {
-            return content;
-        }
-        
-        List<String> tempUrls = extractTempImageUrls(content);
-        if (tempUrls.isEmpty()) {
-            return content;
-        }
-        
-        Map<String, String> urlMappings = new HashMap<>();
-        for (String tempUrl : tempUrls) {
-            try {
-                String permanentUrl = migrateTempImageByUrl(tempUrl, imageType, userId, additionalPath);
-                urlMappings.put(tempUrl, permanentUrl);
-            } catch (Exception e) {
-                log.error("이미지 이동 실패: tempUrl={}, userId={}", tempUrl, userId, e);
-            }
-        }
-        
-        return replaceTempUrls(content, urlMappings);
-    }
-
-    // 본문에서 temp 이미지 URL 패턴 추출
-    private List<String> extractTempImageUrls(String content) {
-        if (tempUrlPattern == null) {
-            String pattern = imageStorageService.getTempImageUrlPattern();
-            log.info("temp URL 정규식 패턴: {}", pattern);
-            tempUrlPattern = Pattern.compile(pattern, Pattern.CASE_INSENSITIVE);
-        }
-        
-        log.info("컨텐츠에서 temp URL 추출 시도: {}", content);
-        
-        Matcher matcher = tempUrlPattern.matcher(content);
-        List<String> tempUrls = new ArrayList<>();
-        while (matcher.find()) {
-            String foundUrl = matcher.group();
-            tempUrls.add(foundUrl);
-            log.info("temp URL 발견: {}", foundUrl);
-        }
-        
-        log.info("총 발견된 temp URL 개수: {}", tempUrls.size());
-        return tempUrls;
-    }
 
     // temp URL을 permanent URL로 이동 처리
     private String migrateTempImageByUrl(String tempUrl, ImageType imageType, Long userId, String additionalPath) {
@@ -296,7 +276,7 @@ public class ImageService {
             throw new BusinessException(ErrorCode.INVALID_INPUT_VALUE);
         }
         
-        if (isExpiredTempImage(tempImage)) {
+        if (isExpiredTempImage(tempImage)){
             throw new BusinessException(ErrorCode.INVALID_INPUT_VALUE);
         }
         
@@ -317,57 +297,12 @@ public class ImageService {
     }
 
 
-    // 본문의 temp URL을 permanent URL로 교체
-    private String replaceTempUrls(String content, Map<String, String> urlMappings) {
-        String result = content;
-        for (Map.Entry<String, String> entry : urlMappings.entrySet()) {
-            result = result.replace(entry.getKey(), entry.getValue());
-        }
-        return result;
-    }
 
-    // temp 경로 이미지인지 확인
-    private boolean isTempImage(String s3Key) {
-        return s3Key != null && s3Key.contains("/temp/");
-    }
 
     // 만료된 temp 이미지인지 확인
     private boolean isExpiredTempImage(Image image) {
-        return isTempImage(image.getS3Key()) && 
+        return imagePathService.isTempImage(image.getS3Key()) && 
             image.getCreatedAt().isBefore(LocalDateTime.now().minusHours(TEMP_IMAGE_EXPIRY_HOURS));
-    }
-
-    // 게시글 수정 시 삭제된 이미지 처리
-    @Transactional
-    public void processDeletedImagesAfterPostUpdate(ImageType imageType, Long postId, String newContent) {
-        // 현재 게시글에 연결된 이미지 목록 조회
-        List<Image> currentImages = imageRepository.findByTypeAndPostId(imageType, postId);
-        
-        if (currentImages.isEmpty()) {
-            return;
-        }
-        
-        // 새 컨텐츠에서 이미지 URL 추출
-        List<String> newImageUrls = extractAllImageUrls(newContent);
-        
-        // 삭제할 이미지 찾기 (현재 DB에는 있지만 새 컨텐츠에는 없는 이미지)
-        List<Image> imagesToDelete = currentImages.stream()
-                .filter(image -> {
-                    String imageUrl = imageStorageService.generatePublicUrl(image.getS3Key());
-                    return !newImageUrls.contains(imageUrl);
-                })
-                .collect(Collectors.toList());
-        
-        // 삭제 처리
-        for (Image imageToDelete : imagesToDelete) {
-            try {
-                deleteImage(imageToDelete.getId());
-                log.info("게시글 수정으로 인한 이미지 삭제: imageId={}, postId={}, imageType={}", 
-                        imageToDelete.getId(), postId, imageType);
-            } catch (Exception e) {
-                log.error("이미지 삭제 실패: imageId={}, postId={}", imageToDelete.getId(), postId, e);
-            }
-        }
     }
     
     // postId로 이미지 목록 조회
@@ -375,26 +310,6 @@ public class ImageService {
         return imageRepository.findByTypeAndPostId(imageType, postId);
     }
     
-    // 컨텐츠에서 모든 이미지 URL 추출 (temp 이미지뿐만 아니라 permanent 이미지도)
-    private List<String> extractAllImageUrls(String content) {
-        if (content == null || content.trim().isEmpty()) {
-            return new ArrayList<>();
-        }
-        
-        String allImagePattern = imageStorageService.getAllImageUrlPattern();
-        Pattern pattern = Pattern.compile(allImagePattern, Pattern.CASE_INSENSITIVE);
-        
-        Matcher matcher = pattern.matcher(content);
-        List<String> imageUrls = new ArrayList<>();
-        while (matcher.find()) {
-            imageUrls.add(matcher.group());
-        }
-        
-        log.debug("컨텐츠에서 추출된 이미지 URL 개수: {}", imageUrls.size());
-        return imageUrls;
-    }
-
-    // 공통 유틸리티 메서드들
     
     // 이미지 메타데이터 추출
     private ImageMetadata extractImageMetadata(MultipartFile file) throws IOException {
@@ -402,42 +317,6 @@ public class ImageService {
         Long width = bufferedImage != null ? (long) bufferedImage.getWidth() : null;
         Long height = bufferedImage != null ? (long) bufferedImage.getHeight() : null;
         return new ImageMetadata(width, height);
-    }
-    
-    // temp 경로 생성
-    private String generateTempPath(Long userId) {
-        String timestamp = LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyyMMdd_HHmmss"));
-        String randomId = UUID.randomUUID().toString().substring(0, 8);
-        return String.format("users/%d/temp/%s_%s", userId, timestamp, randomId);
-    }
-    
-    // 다중 파일 검증
-    private void validateMultipleFiles(List<MultipartFile> files) {
-        if (files == null || files.isEmpty()) {
-            throw new BusinessException(ErrorCode.REQUIRED_FIELD_MISSING);
-        }
-        
-        if (files.size() > MAX_FILES_PER_UPLOAD) {
-            throw new BusinessException(ErrorCode.INVALID_INPUT_VALUE);
-        }
-        
-        files.forEach(this::validateImageFile);
-    }
-    
-    // 단일 파일 검증
-    private void validateImageFile(MultipartFile file) {
-        if (file == null || file.isEmpty()) {
-            throw new BusinessException(ErrorCode.REQUIRED_FIELD_MISSING);
-        }
-        
-        if (file.getSize() > MAX_FILE_SIZE_BYTES) {
-            throw new BusinessException(ErrorCode.IMAGE_SIZE_TOO_LARGE);
-        }
-        
-        String contentType = file.getContentType();
-        if (contentType == null || !contentType.startsWith("image/")) {
-            throw new BusinessException(ErrorCode.IMAGE_FORMAT_NOT_SUPPORTED);
-        }
     }
     
     // 이미지 메타데이터 내부 클래스
